+++ parsers/falcon.c
@@ -52,7 +52,7 @@ static bool isIdentifierChar (int c)
 
 static const unsigned char *skipSpace (const unsigned char *cp)
 {
-    while (isspace ((int) *cp))
+    while (isspace (*cp))
         ++cp;
 
     return cp;
@@ -81,7 +81,7 @@ static void findFalconTags (void)
             cp += 8;
             cp = skipSpace (cp);
 
-            while (isIdentifierChar ((int) *cp))
+            while (isIdentifierChar (*cp))
             {
                 vStringPut (name, *cp);
                 ++cp;
@@ -94,7 +94,7 @@ static void findFalconTags (void)
             cp += 5;
             cp = skipSpace (cp);
 
-            while (isIdentifierChar ((int) *cp))
+            while (isIdentifierChar (*cp))
             {
                 vStringPut (name, *cp);
                 ++cp;
@@ -107,7 +107,7 @@ static void findFalconTags (void)
             cp += 4;
             cp = skipSpace (cp);
 
-            while (isIdentifierChar ((int) *cp))
+            while (isIdentifierChar (*cp))
             {
                 vStringPut (name, *cp);
                 ++cp;
@@ -120,7 +120,7 @@ static void findFalconTags (void)
             cp += 12;
             cp = skipSpace (cp);
 
-            while (isIdentifierChar ((int) *cp))
+            while (isIdentifierChar (*cp))
             {
                 vStringPut (name, *cp);
                 ++cp;
